HIP-1023: Allow each sub-grid to be processed in parallel. Improve w-towers timing report. Other minor tweaks and fixes.

+++ src/ska-sdp-func/grid_data/sdp_gridder_utils.h
@@ -185,7 +185,7 @@ void sdp_gridder_subgrid_cut_out(
 void sdp_gridder_sum_diff(
         const sdp_Mem* a,
         const sdp_Mem* b,
-        double* result,
+        int64_t* result,
         sdp_Error* status
 );
